# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hæstore API",
    "version": "0.2.0",
    "description": "The catalogue is admin-defined: categories, attributes and variants are data, not code. Endpoints that return filters or attribute sets describe whatever an administrator has configured, so a storefront built against this document needs no hardcoded knowledge of any particular product kind."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
...
      "ListingCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "priceRange": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "min": {
                "type": "integer"
              },
              "max": {
                "type": "integer"
              },
              "currency": {
                "type": "string"
              }
            },
            "required": [
              "min",
              "max",
              "currency"
            ]
          },
          "inStock": {
            "type": "boolean"
          },
          "image": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "publicId": {
                "type": "string"
              },
              "alt": {
                "type": "string"
              },
              "width": {
                "type": "integer"
              },
              "height": {
                "type": "integer"
              },
              "blurDataUrl": {
                "type": "string"
              }
            },
            "required": [
              "publicId",
              "alt"
            ]
          },
          "ratingAverage": {
            "type": "number"
          },
          "ratingCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "priceRange",
          "inStock",
          "image",
          "ratingAverage",
          "ratingCount"
        ]
      },
...
    },
...
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hae_sid",
        "description": "Opaque 256-bit session id, looked up in Redis. There are no JWTs anywhere in this system. The __Host- prefix forbids a Domain attribute, so a compromised subdomain cannot set a session cookie for the API."
      }
    }
  },
  "paths": {
...
    "/api/catalog/products": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "The storefront listing: search, filters, sorting and facet counts.",
        "description": "Served by Meilisearch, with a MongoDB fallback behind the same URL. `page.degraded` says which answered: when it is true, `facets` is null and attribute filters were not applied — each one comes back in `ignoredFilters` with a reason.\n\n**Any parameter not listed here is treated as an attribute filter**, matched against the category's own AttributeDefinitions. That is how a filter an admin defined this morning works without a deploy: `?roast=medium,dark&weight_g=250-1000`. Values within one attribute are OR’d; different attributes are AND’d. An unrecognised key or value is reported in `ignoredFilters` rather than rejected, so a bookmark outlives the attribute it names.\n\nPage size defaults to 24 and is capped at 60; depth is capped at 1000 documents.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Full-text query."
            },
            "required": false,
            "description": "Full-text query.",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "coffee-tea/beans"
            },
            "required": false,
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 60
            },
            "required": false,
            "name": "per_page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "relevance",
                "newest",
                "oldest",
                "price_asc",
                "price_desc",
                "rating"
              ],
              "description": "A whitelisted enum, never a raw sort expression. Defaults to relevance with a query and newest without one."
            },
            "required": false,
            "description": "A whitelisted enum, never a raw sort expression. Defaults to relevance with a query and newest without one.",
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "1500-4000",
              "description": "Minor units."
            },
            "required": false,
            "description": "Minor units.",
            "name": "price",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "required": false,
            "name": "in_stock",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of product cards with the generated facet panel.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ListingCard"
                      }
                    },
                    "page": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "perPage": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "degraded": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "page",
                        "perPage",
                        "total",
                        "totalPages",
                        "degraded"
                      ]
                    },
                    "facets": {
                      "type": [
                        "array",
                        "null"
                      ],
                      "items": {
                        "$ref": "#/components/schemas/Facet"
                      }
                    },
                    "ignoredFilters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "key",
                          "reason"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data",
                    "page",
                    "facets",
                    "ignoredFilters"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "The request was malformed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/products/{slug}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "One product, with its attributes and variants.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "slug",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The product.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/catalog/attributes": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "Every attribute definition.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "includeArchived",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Definitions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AttributeDefinition"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "Define a new attribute.",
        "description": "The key is permanent. It becomes a Meilisearch filterable attribute, a public URL parameter, the discriminator on every stored value and an entry in variantAxes, so renaming would break bookmarked filter URLs and orphan the index.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9_]{1,39}$"
                  },
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "select",
                      "multiselect",
                      "text",
                      "number",
                      "boolean",
                      "color",
                      "dimension"
                    ]
                  },
                  "unit": {
                    "type": "string",
                    "maxLength": 16
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 60,
                          "pattern": "^[a-z0-9][a-z0-9_-]*$/i"
                        },
                        "label": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "swatchHex": {
                          "type": "string",
                          "pattern": "^#[0-9a-fA-F]{6}$"
                        },
                        "order": {
                          "type": "integer",
                          "minimum": 0,
                          "default": 0
                        }
                      },
                      "required": [
                        "value",
                        "label"
                      ]
                    },
                    "maxItems": 200,
                    "default": []
                  },
                  "isFilterable": {
                    "type": "boolean",
                    "default": true
                  },
                  "isSearchable": {
                    "type": "boolean",
                    "default": false
                  },
                  "isVariantAxis": {
                    "type": "boolean",
                    "default": false
                  },
                  "filterUi": {
                    "type": "string",
                    "enum": [
                      "checkbox",
                      "swatch",
                      "range",
                      "toggle",
                      "select"
                    ],
                    "default": "checkbox"
                  },
                  "validation": {
                    "type": "object",
                    "properties": {
                      "min": {
                        "type": "number"
                      },
                      "max": {
                        "type": "number"
                      },
                      "step": {
                        "type": "number",
                        "exclusiveMinimum": 0
                      },
                      "maxLength": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 5000
                      },
                      "requiredByDefault": {
                        "type": "boolean",
                        "default": false
                      }
                    },
                    "default": {
                      "requiredByDefault": false
                    }
                  }
                },
                "required": [
                  "key",
                  "label",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AttributeDefinition"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with something that already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/catalog/attributes/{id}": {
      "patch": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "Edit an attribute. Its key and type cannot change.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "unit": {
                    "type": "string",
                    "maxLength": 16
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 60,
                          "pattern": "^[a-z0-9][a-z0-9_-]*$/i"
                        },
                        "label": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "swatchHex": {
                          "type": "string",
                          "pattern": "^#[0-9a-fA-F]{6}$"
                        },
                        "order": {
                          "type": "integer",
                          "minimum": 0,
                          "default": 0
                        }
                      },
                      "required": [
                        "value",
                        "label"
                      ]
                    },
                    "maxItems": 200
                  },
                  "isFilterable": {
                    "type": "boolean"
                  },
                  "isSearchable": {
                    "type": "boolean"
                  },
                  "isVariantAxis": {
                    "type": "boolean"
                  },
                  "filterUi": {
                    "type": "string",
                    "enum": [
                      "checkbox",
                      "swatch",
                      "range",
                      "toggle",
                      "select"
                    ]
                  },
                  "validation": {
                    "type": "object",
                    "properties": {
                      "min": {
                        "type": "number"
                      },
                      "max": {
                        "type": "number"
                      },
                      "step": {
                        "type": "number",
                        "exclusiveMinimum": 0
                      },
                      "maxLength": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 5000
                      },
                      "requiredByDefault": {
                        "type": "boolean",
                        "default": false
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AttributeDefinition"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "One attribute definition.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The definition.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AttributeDefinition"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/catalog/categories/{id}/effective-attributes": {
      "get": {
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "tags": [
          "Admin catalogue"
        ],
        "summary": "Every attribute that applies here, inherited ones included.",
        "description": "What the product form renders from. Each says which ancestor contributed it.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The effective set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EffectiveAttributeSet"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not visible to this caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}